backend/whatsapp/requirements.txt
selenium==4.15.2
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
webdriver-manager==4.0.1
pillow==10.1.0
cryptography==41.0.7
python-dotenv==1.0.0
